move Slogra & Gaibon to it's own file
